requirements.txt
Flask==3.1.3
Flask-SocketIO==5.6.1
Flask-Limiter==4.1.0
eventlet==0.41.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
Werkzeug==3.1.3
gunicorn==22.0.0
